Removing `struct_names(true)` from `PrettyConfig`
Including the struct names for serialization creates problems
if you were to change the name of the struct in code.

If you, for example, have a struct `Config`, `store` it, and change
the name of the struct to `AppConfig`, you would get a deserialization
error because the deserializer is expecting the old name of the struct.